Calculate the users score
When the user starts a level a document is written to the userTimestamps
collection. It contains a serverTimestamp() of when they started. When the
user completes the level the document is updated with a serverTimestamp()
for the end time. With these two values a totalTime is calculated and saved
in the same document. This document represents the users gameplay session.