Add functional leaderboards
After completing a stage the user is prompted to enter their name.
Their score is then looked up off the server using their session id. A
new entry is added to the leaderboard collection containing their name
and their score. The db is then queried for the top 10 scores and they are
displayed.