Correctly render reticles on window resize
ReticlePositions now store their position normalized to the image, not the page.
When the positions are mapped to img elements they take into account the
images top and left offset.

Initially I had a ref to the image in a useRef. However on intial page
load the offsets were not accurate. After a page resize they worked fine.
Now I gave the image element a callback ref. This successfully get's
triggered after the image is fully loaded and the offset positions are correct.
I then set a variable to the element passed to this callback, and this element
is referenced in the window resize callback and offset's are working correctly.